requirements.txt
beautifulsoup4==4.10.0
flask==2.1.1
openai==0.27.0
Werkzeug==2.2.2
schedule==1.2.1
tiktoken==0.5.2
python-dotenv==1.0.0
